% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggx.R
\name{gg_}
\alias{gg_}
\title{Generic command to transform a query in natural language
into a ggplot2 command.}
\usage{
gg_(wish = NULL, ...)
}
\arguments{
\item{wish}{Character.}
}
\description{
Generic command to transform a query in natural language
into a ggplot2 command.
}
