% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggx.R
\name{gghelp}
\alias{gghelp}
\title{Helps you find a ggplot command for formatting the plot.}
\usage{
gghelp(wish = "", print = TRUE)
}
\arguments{
\item{wish}{Character. A natural language command to ggplot2.}

\item{print}{Boolean. Print out the command or just return it.}
}
\description{
Helps you find a ggplot command for formatting the plot.
}
